Kreiranje jsp stranice za potvrdu odabranih mesta i kreiranje metoda za odabir mesta u sali @T2568:R @T2567:R

+++ BioskopJPA/src/model/Sifarnik.java
@@ -19,7 +19,7 @@ public class Sifarnik implements Serializable {
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	private int sifarnikID;
 
-	private BigDecimal cena;
+	private double cena;
 
 	private String tip;
 
@@ -38,11 +38,11 @@ public class Sifarnik implements Serializable {
 		this.sifarnikID = sifarnikID;
 	}
 
-	public BigDecimal getCena() {
+	public double getCena() {
 		return this.cena;
 	}
 
-	public void setCena(BigDecimal cena) {
+	public void setCena(double cena) {
 		this.cena = cena;
 	}
